# api-doc/swagger-ui/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Sistem Informasi Web Gereja Kanak-kanak Yesus Marau",
    "description": "Sistem Informasi Web Gereja Kanak-kanak Yesus Marau",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "Sistem Informasi Web Gereja Kanak-kanak Yesus Marau Logo",
      "altText": "Sistem Informasi Web Gereja Kanak-kanak Yesus Marau Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "Sistem Informasi Web Gereja Kanak-kanak Yesus Marau API Support",
      "url": "https://gerejamarau.com/info/support",
      "email": "support@gerejamarau.com"
    },
    "termsOfService": "https://gerejamarau.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/account": {
      "get": {
        "summary": "User Account Details",
        "description": "My Account",
        "operationId": "Account",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Account"
        ],
        
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Account": {
        "type": "object",
        "properties": {
          "user_id": {
            "description": "User Id",
            "type": "string"
          },
          "user_name": {
            "description": "User Name",
            "type": "string"
          },
          "email": {
            "description": "Email",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "user_role_id": {
            "description": "User Role Id",
            "type": "string"
          },
          "account_status": {
            "description": "Account Status",
            "type": "string"
          },
          "email_verified_at": {
            "description": "Email Verified At",
            "type": "string"
          }
        }
      },
...
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}